Lesson5Team Test Assignment

+++ src/nephisJourney/src/control/AltarControl.java
@@ -12,7 +12,7 @@ package nephisJourney.src.control;
 public class AltarControl {
 
 //modifiers returnType functionName (datatype paramerer1, ...)
-    public double calcAltarSize(double height, double width) { 
+    public int calcAltarSize(int height, int width) { 
         if (height < 1 || height > 4){ //test to check if the height is in range
 		return -1;
         }
@@ -20,7 +20,7 @@ public class AltarControl {
 		return -2;
         }
         
-        double radius = width / 2; //divides the width to give the radius
+        double radius = width / 2.0; //divides the width to give the radius
         int volume =(int) (Math.PI * Math.pow(radius, 2) * height);
         /*finds the volume of the Altar and return it as a whole number of 
         stones needed
